Remove obsolete and confusing module layer
As it turns out just simple extending of the module does already the magic :)
